feat(death): Upgraded death detection (for thorns).

+++ src/main/java/fr/hashtek/spigot/breakffa/listener/ListenerDamage.java
@@ -5,6 +5,7 @@ import fr.hashtek.spigot.breakffa.death.DeathReason;
 import fr.hashtek.spigot.breakffa.game.GameManager;
 import fr.hashtek.spigot.breakffa.player.PlayerData;
 import fr.hashtek.spigot.breakffa.player.PlayerManager;
+import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -30,6 +31,9 @@ public class ListenerDamage implements Listener
     /**
      * Called when an entity takes damage.
      * In this case, we only process players.
+     * FIXME: Detection only applies to chestplates, as the
+     *        only armor piece that will have Thorns enchantment
+     *        are chestplates.
      */
     @EventHandler
     public void onDamage(EntityDamageEvent event)
@@ -39,8 +43,10 @@ public class ListenerDamage implements Listener
 
         /* If damage cause is unwanted, cancel the event. */
         if (event.getCause() == EntityDamageEvent.DamageCause.FALL ||
-            event.getCause() == EntityDamageEvent.DamageCause.ENTITY_EXPLOSION)
+            event.getCause() == EntityDamageEvent.DamageCause.ENTITY_EXPLOSION) {
             event.setCancelled(true);
+            return;
+        }
 
         final Player player = (Player) event.getEntity();
         final PlayerData playerData = this.gameManager.getPlayerData(player);
@@ -48,6 +54,12 @@ public class ListenerDamage implements Listener
 
         /* If player is going to die, custom kill. */
         if (player.getHealth() - event.getFinalDamage() <= 0) {
+            if (event.getCause() == EntityDamageEvent.DamageCause.ENTITY_ATTACK)
+                return;
+
+            if (event.getCause() == EntityDamageEvent.DamageCause.THORNS)
+                playerData.setLastDamagerWeapon(playerData.getLastDamager().getInventory().getChestplate());
+
             event.setCancelled(true);
             playerManager.kill(DeathReason.fromDamageCause(event.getCause()));
         }
